DI Task: - some refactor without change of logic
