+++ app/src/main/java/com/example/todolist/di/ViewModelModule.kt
@@ -1,8 +1,8 @@
 package com.example.todolist.di
 
 import androidx.lifecycle.ViewModel
-import com.example.todolist.presantation.taskItem.TaskItemViewModel
-import com.example.todolist.presantation.toDoList.ToDoListViewModel
+import com.example.todolist.presantation.taskItemFragment.TaskItemViewModel
+import com.example.todolist.presantation.toDoListFragment.ToDoListViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
