+++ app/src/main/java/com/example/todolist/domain/useCases/UpgradeTaskItemUseCase.kt
@@ -1,5 +1,7 @@
-package com.example.todolist.domain
+package com.example.todolist.domain.useCases
 
+import com.example.todolist.domain.Repository
+import com.example.todolist.domain.TaskItem
 import javax.inject.Inject
 
 class UpgradeTaskItemUseCase @Inject constructor(private val repository: Repository) {
