+++ app/src/main/java/com/example/todolist/presantation/MainActivity.kt
@@ -3,7 +3,7 @@ package com.example.todolist.presantation
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.todolist.R
-import com.example.todolist.presantation.taskItem.TaskItemFragment
+import com.example.todolist.presantation.taskItemFragment.TaskItemFragment
 
 class MainActivity : AppCompatActivity(), TaskItemFragment.OnEditingIsFinishedListener {
     override fun onCreate(savedInstanceState: Bundle?) {
