+++ app/src/main/java/com/example/todolist/presantation/taskItemFragment/TaskItemViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.todolist.presantation.taskItem
+package com.example.todolist.presantation.taskItemFragment
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.todolist.domain.AddTaskItemUseCase
-import com.example.todolist.domain.GetTaskItemUseCase
+import com.example.todolist.domain.useCases.AddTaskItemUseCase
+import com.example.todolist.domain.useCases.GetTaskItemUseCase
 import com.example.todolist.domain.TaskItem
-import com.example.todolist.domain.UpgradeTaskItemUseCase
+import com.example.todolist.domain.useCases.UpgradeTaskItemUseCase
 import kotlinx.coroutines.launch
 import javax.inject.Inject
