+++ app/src/main/java/com/example/todolist/presantation/toDoListFragment/customDailyCalendarViewGroup/CustomDailyCalendarViewGroup.kt
@@ -1,4 +1,4 @@
-package com.example.todolist.presantation.toDoList.customDailyCalendarViewGroup
+package com.example.todolist.presantation.toDoListFragment.customDailyCalendarViewGroup
 
 import android.animation.LayoutTransition
 import android.annotation.SuppressLint
