+++ app/src/main/res/layout/fragment_to_do_list.xml
@@ -7,7 +7,7 @@
     android:layout_margin="16dp"
     android:elevation="8dp"
     app:cardCornerRadius="30dp"
-    tools:context=".presantation.toDoList.ToDoListFragment">
+    tools:context=".presantation.toDoListFragment.ToDoListFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
@@ -69,14 +69,14 @@
                 android:contentDescription="@string/container_for_daily_calendar"
                 >
 
-                <com.example.todolist.presantation.toDoList.customDailyCalendarViewGroup.CustomDailyCalendarViewGroup
+                <com.example.todolist.presantation.toDoListFragment.customDailyCalendarViewGroup.CustomDailyCalendarViewGroup
                     android:id="@+id/daily_calendar_view_group"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     app:eachRowHeight="20dp"
                     app:minimumPercentage="0.25"
                     app:numberOfRows="24">
-                </com.example.todolist.presantation.toDoList.customDailyCalendarViewGroup.CustomDailyCalendarViewGroup>
+                </com.example.todolist.presantation.toDoListFragment.customDailyCalendarViewGroup.CustomDailyCalendarViewGroup>
             </HorizontalScrollView>
         </ScrollView>
